Refactor: Extract declared method into separate DSL module
- Extract declared method logic from InsideRoute::PostBeforeFilter into new Declared module
- Simplify InsideRoute by replacing dynamic module extension with standard include
- Add before_filter_passed flag in Endpoint to track filter state
- Maintain backward compatibility by aliasing MethodNotYetAvailable exception
- Update tests to reflect new module structure
- Add CHANGELOG entry

This refactoring improves code organization and maintainability while
preserving all existing functionality and API compatibility.
